+++ src/drivers/imu/bosch/bmi088/BMI088_Accelerometer.cpp
@@ -220,7 +220,7 @@ void BMI088_Accelerometer::RunImpl()
 
 					// tolerate minor jitter, leave sample to next iteration if behind by only 1
 					if (samples == _fifo_samples + 1) {
-						timestamp_sample -= FIFO_SAMPLE_DT;
+						timestamp_sample -= static_cast<int>(FIFO_SAMPLE_DT);
 						samples--;
 					}
